// src/core/api/generated/spring/schemas/ProductImageDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export interface ProductImageDTO {
  id?: number;
  productId: number;
  /**
   * @minLength 0
   * @maxLength 255
   */
  gumletAssetId: string;
  /**
   * @minLength 0
   * @maxLength 500
   */
  gumletPath: string;
  displayOrder: number;
  /**
   * @minLength 0
   * @maxLength 255
   */
  originalFilename: string;
  /**
   * @minLength 0
   * @maxLength 20
   */
  format: string;
  fileSizeBytes: number;
  width?: number;
  height?: number;
  organizationId: number;
  deletedAt?: string;
  createdBy?: string;
  createdDate?: string;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  cdnUrl?: string;
  thumbnailUrl?: string;
  isPrimary?: boolean;
}